use Canary from duty models
