+++ api/v1/canary_types.go
@@ -24,6 +24,8 @@ import (
 
 	"github.com/flanksource/canary-checker/api/external"
 	"github.com/flanksource/commons/logger"
+	"github.com/flanksource/duty/models"
+	"github.com/flanksource/duty/types"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
@@ -306,6 +308,52 @@ type Canary struct {
 
 	Spec   CanarySpec   `json:"spec,omitempty"`
 	Status CanaryStatus `json:"status,omitempty"`
+}
+
+func (canary Canary) ToModel() (models.Canary, error) {
+	spec, err := json.Marshal(canary.Spec)
+	if err != nil {
+		return models.Canary{}, err
+	}
+	var checks = make(map[string]string)
+	if canary.Status.Checks != nil {
+		checks = canary.Status.Checks
+	}
+
+	return models.Canary{
+		Spec:      spec,
+		Labels:    types.JSONStringMap(canary.Labels),
+		Name:      canary.Name,
+		Namespace: canary.Namespace,
+		Source:    canary.Annotations["source"],
+		Checks:    types.JSONStringMap(checks),
+	}, nil
+}
+
+func CanaryFromModel(c *models.Canary) (*Canary, error) {
+	canary := Canary{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      c.Name,
+			Namespace: c.Namespace,
+			Annotations: map[string]string{
+				"source": c.Source,
+			},
+			Labels: c.Labels,
+		},
+	}
+	var deletionTimestamp metav1.Time
+	if c.DeletedAt != nil && !c.DeletedAt.IsZero() {
+		deletionTimestamp = metav1.NewTime(*c.DeletedAt)
+		canary.ObjectMeta.DeletionTimestamp = &deletionTimestamp
+	}
+	if err := json.Unmarshal(c.Spec, &canary.Spec); err != nil {
+		logger.Debugf("Failed to unmarshal canary spec: %s", err)
+		return nil, err
+	}
+	id := c.ID.String()
+	canary.Status.PersistedID = &id
+	canary.Status.Checks = c.Checks
+	return &canary, nil
 }
 
 func NewCanaryFromSpec(name string, spec CanarySpec) Canary {
