+++ api/v1/db_types.go
@@ -16,7 +16,6 @@ const (
 	SQLServerType = "sqlserver"
 	PostgresType  = "postgres"
 	SqliteType    = "sqlite"
-	text          = "TEXT"
 	jsonType      = "json"
 	jsonbType     = "JSONB"
 	nvarcharType  = "NVARCHAR(MAX)"
