+++ pkg/db/canary.go
@@ -9,19 +9,19 @@ import (
 
 	v1 "github.com/flanksource/canary-checker/api/v1"
 	"github.com/flanksource/canary-checker/pkg"
-	"github.com/flanksource/canary-checker/pkg/db/types"
 	"github.com/flanksource/canary-checker/pkg/metrics"
 	"github.com/flanksource/canary-checker/pkg/utils"
 	"github.com/flanksource/commons/logger"
 	"github.com/flanksource/duty"
 	"github.com/flanksource/duty/models"
+	dutyTypes "github.com/flanksource/duty/types"
 	"github.com/google/uuid"
 	"gorm.io/gorm"
 	"gorm.io/gorm/clause"
 )
 
-func GetAllCanaries() ([]pkg.Canary, error) {
-	var _canaries []pkg.Canary
+func GetAllCanaries() ([]models.Canary, error) {
+	var _canaries []models.Canary
 	var rawCanaries interface{}
 	query := fmt.Sprintf("SELECT json_agg(jsonb_set_lax(to_jsonb(canaries),'{checks}', %s)) :: jsonb as canaries from canaries where deleted_at is null", getChecksForCanaries())
 
@@ -115,7 +115,7 @@ func UpdateChecksStatus(ids []string, status models.CheckHealthStatus) error {
 
 func DeleteCanary(canary v1.Canary) error {
 	logger.Infof("deleting canary %s/%s", canary.Namespace, canary.Name)
-	model, err := pkg.CanaryFromV1(canary)
+	model, err := canary.ToModel()
 	if err != nil {
 		return err
 	}
@@ -154,8 +154,8 @@ func DeleteChecks(id []string) error {
 	return Gorm.Table("checks").Where("id IN (?)", id).UpdateColumn("deleted_at", time.Now()).Error
 }
 
-func GetCanary(id string) (*pkg.Canary, error) {
-	var model *pkg.Canary
+func GetCanary(id string) (*models.Canary, error) {
+	var model *models.Canary
 	if err := Gorm.Where("id = ?", id).First(&model).Error; err != nil {
 		return nil, err
 	}
@@ -163,8 +163,8 @@ func GetCanary(id string) (*pkg.Canary, error) {
 	return model, nil
 }
 
-func FindCanaryByID(id string) (*pkg.Canary, error) {
-	var model *pkg.Canary
+func FindCanaryByID(id string) (*models.Canary, error) {
+	var model *models.Canary
 	if err := Gorm.Where("id = ?", id).First(&model).Error; err != nil {
 		if errors.Is(err, gorm.ErrRecordNotFound) {
 			return nil, nil
@@ -187,8 +187,8 @@ func GetCheck(id string) (*pkg.Check, error) {
 	return model, nil
 }
 
-func FindCanary(namespace, name string) (*pkg.Canary, error) {
-	var model pkg.Canary
+func FindCanary(namespace, name string) (*models.Canary, error) {
+	var model models.Canary
 	if err := Gorm.Where("namespace = ? AND name = ?", namespace, name).First(&model).Error; err != nil {
 		if errors.Is(err, gorm.ErrRecordNotFound) {
 			return nil, nil
@@ -199,7 +199,7 @@ func FindCanary(namespace, name string) (*pkg.Canary, error) {
 	return &model, nil
 }
 
-func FindCheck(canary pkg.Canary, name string) (*pkg.Check, error) {
+func FindCheck(canary models.Canary, name string) (*pkg.Check, error) {
 	var model pkg.Check
 	if err := Gorm.Where("canary_id = ? AND name = ?", canary.ID.String(), name).First(&model).Error; err != nil {
 		if errors.Is(err, gorm.ErrRecordNotFound) {
@@ -210,22 +210,22 @@ func FindCheck(canary pkg.Canary, name string) (*pkg.Check, error) {
 	return &model, nil
 }
 
-func CreateCanary(canary *pkg.Canary) error {
+func CreateCanary(canary *models.Canary) error {
 	if canary.Spec == nil || len(canary.Spec) == 0 {
 		empty := []byte("{}")
-		canary.Spec = types.JSON(empty)
+		canary.Spec = dutyTypes.JSON(empty)
 	}
 
 	return Gorm.Create(canary).Error
 }
 
-func CreateCheck(canary pkg.Canary, check *pkg.Check) error {
+func CreateCheck(canary models.Canary, check *pkg.Check) error {
 	return Gorm.Create(&check).Error
 }
 
-func PersistCanary(canary v1.Canary, source string) (*pkg.Canary, map[string]string, bool, error) {
+func PersistCanary(canary v1.Canary, source string) (*models.Canary, map[string]string, bool, error) {
 	changed := false
-	model, err := pkg.CanaryFromV1(canary)
+	model, err := canary.ToModel()
 	if err != nil {
 		return nil, nil, changed, err
 	}
